LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/media-elements/interfaces/TextTrack/kind-expected.txt


PASS TextTrack.kind, addTextTrack 
PASS TextTrack.kind, track element 
FAIL TextTrack.kind, \u0000 assert_equals: expected "metadata" but got "subtitles"
